Use logging everywhere Switch pattern to includes and excludes patterns lists Modify the "changes" computation, make sure the right action takes precedence

+++ localdb.py
@@ -5,6 +5,7 @@ import hashlib
 import time
 import fnmatch
 import pickle
+import logging
 
 from utils import hashfile
 
@@ -127,9 +128,20 @@ class LocalDbHandler():
         return hashfile(self.base + node_path, hashlib.md5())
 
 
+    def compare_raw_pathes(self, row1, row2):
+        if row1['source'] != 'NULL':
+            cmp1 = row1['source']
+        else:
+            cmp1 = row1['target']
+        if row2['source'] != 'NULL':
+            cmp2 = row2['source']
+        else:
+            cmp2 = row2['target']
+        return cmp1 == cmp2
+
     def get_local_changes(self, seq_id, accumulator):
 
-        print "Local sequence " + str(seq_id)
+        logging.debug("Local sequence " + str(seq_id))
         last = seq_id
         conn = sqlite3.connect(self.db)
         conn.row_factory = sqlite3.Row
@@ -139,8 +151,8 @@ class LocalDbHandler():
         orders = dict()
         orders['path'] = 0
         orders['content'] = 1
-        orders['create'] = 2
         orders['delete'] = 3
+        orders['create'] = 3
         for row in c.execute("SELECT * FROM ajxp_changes LEFT JOIN ajxp_index "
                              "ON ajxp_changes.node_id = ajxp_index.node_id "
                              "WHERE seq > ? ORDER BY ajxp_changes.node_id, seq ASC", (seq_id,)):
@@ -149,11 +161,22 @@ class LocalDbHandler():
             for att in ('mtime', 'md5', 'bytesize', 'node_path',):
                 drow['node'][att] = row[att]
                 drow.pop(att, None)
-            if drow['node_id'] == previous_node_id:
+            if drow['node_id'] == previous_node_id and self.compare_raw_pathes(drow, previous_row):
                 previous_row['target'] = drow['target']
                 previous_row['seq'] = drow['seq']
-                if orders[drow['type']] > orders[previous_row['type']]:
+                #if orders[drow['type']] > orders[previous_row['type']]:
+                if (drow['type'] == 'path' or drow['type'] == 'content'):
+                    if previous_row['type'] == 'delete':
+                        previous_row['type'] = drow['type']
+                    elif previous_row['type'] == 'create':
+                        previous_row['type'] = 'create'
+                    else:
+                        previous_row['type'] = drow['type']
+                elif drow['type'] == 'create':
+                    previous_row['type'] = 'create'
+                else:
                     previous_row['type'] = drow['type']
+
             else:
                 if previous_row is not None and (previous_row['source'] != previous_row['target'] or previous_row['type'] == 'content'):
                     previous_row['location'] = 'local'
@@ -166,25 +189,42 @@ class LocalDbHandler():
             previous_row['location'] = 'local'
             accumulator.append(previous_row)
 
+        #refilter: create + delete or delete + create must be ignored
+        for row in accumulator:
+            print str(row['seq']) + '-' + row['type'] + '-' + row['source'] + '-' + row['target']
+
         conn.close()
         return last
 
 
 class SqlEventHandler(FileSystemEventHandler):
 
-    def __init__(self, pattern='*', basepath=''):
+    def __init__(self, basepath, includes, excludes):
         super(SqlEventHandler, self).__init__()
         self.base = basepath
-        self.pattern = pattern
+        self.includes = includes
+        self.excludes = excludes
         db_handler = LocalDbHandler(basepath)
         self.db = db_handler.db
 
     def remove_prefix(self, text):
         return text[len(self.base):] if text.startswith(self.base) else text
 
+    def included(self, event, base=None):
+        if not base:
+            base = os.path.basename(event.src_path)
+        for i in self.includes:
+            if not fnmatch.fnmatch(base, i):
+                return False
+        for e in self.excludes:
+            if fnmatch.fnmatch(base, e):
+                return False
+        return True
+
     def on_moved(self, event):
-        print(event)
-        print("Move noticed: " + event.event_type + " on file " + event.src_path + " at " + time.asctime())
+        logging.debug("Event: move noticed: " + event.event_type + " on file " + event.src_path + " at " + time.asctime())
+        if not self.included(event):
+            return
         conn = sqlite3.connect(self.db)
         t = (
             self.remove_prefix(event.dest_path.decode('utf-8')),
@@ -195,8 +235,11 @@ class SqlEventHandler(FileSystemEventHandler):
         conn.close()
 
     def on_created(self, event):
-        print("Creation noticed: " + event.event_type +
+        logging.debug("Event: creation noticed: " + event.event_type +
                          " on file " + event.src_path + " at " + time.asctime())
+        if not self.included(event):
+            return
+
         search_key = self.remove_prefix(event.src_path.decode('utf-8'))
         if event.is_directory:
             hash_key = 'directory'
@@ -233,8 +276,11 @@ class SqlEventHandler(FileSystemEventHandler):
         conn.close()
 
     def on_deleted(self, event):
-        print("Deletion noticed: " + event.event_type +
+        logging.debug("Event: deletion noticed: " + event.event_type +
                          " on file " + event.src_path + " at " + time.asctime())
+        if not self.included(event):
+            return
+
         conn = sqlite3.connect(self.db)
         conn.execute("DELETE FROM ajxp_index WHERE node_path LIKE ?", (self.remove_prefix(event.src_path.decode('utf-8')) + '%',))
         conn.commit()
@@ -242,34 +288,36 @@ class SqlEventHandler(FileSystemEventHandler):
 
     def on_modified(self, event):
         super(SqlEventHandler, self).on_modified(event)
+        if not self.included(event):
+            return
 
         if event.is_directory:
             files_in_dir = [event.src_path+"/"+f for f in os.listdir(event.src_path)]
             if len(files_in_dir) > 0:
-                modifiedFilename = max(files_in_dir, key=os.path.getmtime)
+                modified_filename = max(files_in_dir, key=os.path.getmtime)
             else:
                 return
-            if os.path.isfile(modifiedFilename) and fnmatch.fnmatch(os.path.basename(modifiedFilename), self.pattern):
-                print "Modified file : %s" % modifiedFilename
+            if os.path.isfile(modified_filename) and self.included(event=None, base=modified_filename):
+                logging.debug("Event: modified file : %s" % modified_filename)
                 conn = sqlite3.connect(self.db)
-                size = os.path.getsize(modifiedFilename)
-                the_hash = hashfile(open(modifiedFilename, 'rb'), hashlib.md5())
-                mtime = os.path.getmtime(modifiedFilename)
-                search_path = self.remove_prefix(modifiedFilename.decode('utf-8'))
-                stat_result = pickle.dumps(os.stat(modifiedFilename))
+                size = os.path.getsize(modified_filename)
+                the_hash = hashfile(open(modified_filename, 'rb'), hashlib.md5())
+                mtime = os.path.getmtime(modified_filename)
+                search_path = self.remove_prefix(modified_filename.decode('utf-8'))
+                stat_result = pickle.dumps(os.stat(modified_filename))
                 t = (size, the_hash, mtime, search_path, mtime, the_hash, stat_result)
                 conn.execute("UPDATE ajxp_index SET bytesize=?, md5=?, mtime=?, stat_result=? WHERE node_path=? AND bytesize!=? AND md5!=?", t)
                 conn.commit()
                 conn.close()
         else:
-            modifiedFilename = event.src_path
-            print "Modified file : %s" % self.remove_prefix(modifiedFilename)
+            modified_filename = event.src_path
+            logging.debug("Event: modified file : %s" % self.remove_prefix(modified_filename))
             conn = sqlite3.connect(self.db)
-            size = os.path.getsize(modifiedFilename)
-            the_hash = hashfile(open(modifiedFilename, 'rb'), hashlib.md5())
-            mtime = os.path.getmtime(modifiedFilename)
-            search_path = self.remove_prefix(modifiedFilename.decode('utf-8'))
-            stat_result = pickle.dumps(os.stat(modifiedFilename))
+            size = os.path.getsize(modified_filename)
+            the_hash = hashfile(open(modified_filename, 'rb'), hashlib.md5())
+            mtime = os.path.getmtime(modified_filename)
+            search_path = self.remove_prefix(modified_filename.decode('utf-8'))
+            stat_result = pickle.dumps(os.stat(modified_filename))
             t = (size, the_hash, mtime, search_path, mtime, the_hash, stat_result)
             conn.execute("UPDATE ajxp_index SET bytesize=?, md5=?, mtime=?, stat_result=? WHERE node_path=? AND bytesize!=? AND md5!=?", t)
             conn.commit()
